+++ app/reads/[readId]/page.js
@@ -1,160 +1,62 @@
-'use client';
-
-import { useEffect, useMemo } from 'react';
-import { Container, Grid, Skeleton, Typography } from '@mui/material';
-import { useQuery, gql } from '@apollo/client';
-import Image from 'next/image';
-import { useParams, useRouter } from 'next/navigation';
-import ImageGallery from 'react-image-gallery';
-import {
-  Menu,
-  Footer,
-  Spacing,
-  ScrollTopBtn,
-  SkeletonReadLoading,
-} from '../../components';
-import { formatDisplayFullDate, genKeyWords } from '../../utils';
-import useStyles from './styles';
+import { ARTICLE_QUERY } from './queries';
 import 'react-image-gallery/styles/css/image-gallery.css';
+import ReadDetail from './ReadDetail';
 
-const ARTICLE_QUERY = gql`
-  query Article($id: ID) {
-    article(where: { id: $id }) {
-      author
-      id
-      metaTags
-      name
-      brief
-      publishedAt
-      content {
-        html
-      }
-      gallery(first: 30) {
-        url(
-          transformation: {
-            image: { resize: { fit: clip, height: 600, width: 970 } }
-          }
-        )
-      }
-      photo {
-        url(
-          transformation: {
-            image: { resize: { fit: clip, height: 600, width: 970 } }
-          }
-        )
-      }
+export async function generateMetadata({ params }) {
+  const { readId } = params;
+  
+  try {
+    const { data } = await fetch(process.env.NEXT_PUBLIC_GRAPHQL_URL, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        query: ARTICLE_QUERY,
+        variables: { id: readId },
+      }),
+    }).then(res => res.json());
+
+    const article = data?.article;
+
+    if (!article) {
+      return {
+        title: 'Article Not Found',
+        description: 'The requested article could not be found.',
+      };
     }
+
+    return {
+      title: `${article.name} - Noise Hanoi`,
+      description: article.brief,
+      openGraph: {
+        title: article.name,
+        description: article.brief,
+        images: article.photo?.url ? [
+          {
+            url: article.photo.url,
+            width: 970,
+            height: 600,
+            alt: article.name,
+          },
+        ] : [],
+      },
+      twitter: {
+        card: 'summary_large_image',
+        title: article.name,
+        description: article.brief,
+        images: article.photo?.url ? [article.photo.url] : [],
+      },
+    };
+  } catch (error) {
+    console.error('Error generating metadata:', error);
+    return {
+      title: 'Noise Hanoi',
+      description: 'Live Music in Hanoi',
+    };
   }
-`;
-
-const genImages = (arr) => {
-  if (!arr || arr.length === 0) return null;
-  return arr.map((g) => ({
-    original: g.url,
-    thumbnail: g.url,
-  }));
-};
-
-function ReadDetail() {
-  const classes = useStyles();
-  const { readId } = useParams();
-  const router = useRouter();
-  const { data, loading } = useQuery(ARTICLE_QUERY, {
-    variables: {
-      id: readId,
-    },
-  });
-  const articleData = data?.article ?? null;
-  const imageList = useMemo(
-    () => genImages(articleData && articleData.gallery),
-    [articleData]
-  );
-  const keyWords = useMemo(
-    () => genKeyWords(articleData && articleData.metaTags),
-    [articleData]
-  );
-
-  useEffect(() => {
-    window.scrollTo({
-      top: 0,
-      behavior: 'smooth',
-    });
-  }, []);
-
-  if (!loading && !articleData) {
-    router.push('/not-found');
-  }
-  return (
-    <Container maxWidth="lg">
-      <Menu />
-      <Spacing size={48} />
-      <div className={classes.main}>
-        <Grid container spacing={4}>
-          <Grid item xs={12} md={9}>
-            {loading && <SkeletonReadLoading length={4} />}
-            {!loading && articleData && (
-              <>
-                <div className={classes.readName}>{articleData.name}</div>
-                {articleData?.photo?.url && (
-                  <Image
-                    src={articleData?.photo?.url}
-                    alt="something went wrong"
-                    className={classes.readImg}
-                    width={200}
-                    height={200}
-                    style={{
-                      width: '100%',
-                      height: 'auto',
-                    }}
-                  />
-                )}
-                <Spacing size={32} />
-                <div
-                  className={classes.text}
-                  dangerouslySetInnerHTML={{
-                    __html: articleData?.content?.html ?? '',
-                  }}
-                />
-              </>
-            )}
-            <Spacing size={32} />
-            {imageList && <ImageGallery items={imageList} />}
-          </Grid>
-          <Grid item xs={12} md={3}>
-            {loading && (
-              <>
-                <Spacing size={12} />
-                <Typography component="div" key="h5" variant="h5">
-                  <Skeleton width="50%" animation="false" />
-                </Typography>
-                <Spacing size={12} />
-                <Skeleton width="80%" animation="wave" />
-                <Spacing size={32} />
-                <Typography component="div" key="h5" variant="h5">
-                  <Skeleton width="50%" animation="false" />
-                </Typography>
-                <Spacing size={12} />
-                <Skeleton width="80%" animation="wave" />
-              </>
-            )}
-            {!loading && articleData && (
-              <>
-                <div className={classes.label}>By</div>
-                <div className={classes.value}>{articleData?.author}</div>
-                <Spacing size={32} />
-                <div className={classes.label}>Published</div>
-                <div className={classes.value}>
-                  {formatDisplayFullDate(articleData?.publishedAt)}
-                </div>
-              </>
-            )}
-          </Grid>
-        </Grid>
-      </div>
-      <Footer />
-      <ScrollTopBtn />
-    </Container>
-  );
 }
 
-export default ReadDetail;
+export default function Page({ params }) {
+  return <ReadDetail readId={params.readId} />;
+}
